# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "WH40K API",
    "contact": {
      "email": "grineldosanchez@yahoo.es",
      "name": "Admin",
      "url": "http://relatosdelaherejia.cl"
    },
    "description": "Api de servicio WH40K. Para torneos OT500"
  },
  "tags": [
    {
    "name" : "roster-api",
    "description" : "All Roster Battlescribe Services"
    },
    {
      "name": "challonge-tournaments",
      "description": "Servicios de challonge API para torneos"
    },
    {
      "name": "army-api",
      "description": "Army, Factions, SubFactions API"
    },
    {
      "name" : "wahapedia-api",
      "description" : "Warhammer 40K Information, Factions, Rules, Stratagems"
    }  
  ],
  "paths": {
    "/roster/save" : {
      "post" : {
        "tags" : [ "roster-api" ],
        "summary" : "Save Roster",
        "description" : "Guarda el Roster de Battlescribe en Mongodb",
        "operationId" : "saveRoster",
        "consumes" : [ "multipart/form-data" ],
        "produces" : [ "application/json" ],
        "parameters" : [ 
          {
          "name" : "roster_json",
          "in" : "formData",
          "description" : "Json Roster here.",
          "required" : true,
          "type" : "string",
          "default" : "{status:'VALIDATE', name:'Astra Millitarum-OT500-Primaris Psyker'}"
          },
          {
            "name" : "roster_file",
            "in" : "formData",
            "description" : "File Roster here. (ros extension)",
            "required" : true,
            "type" : "file"          
            }
          ],
        "responses" : {
          "200" : {
            "description" : "Resultado de la lista de tournaments[]",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Roster"
              }
            }
          },
          "400" : {
            "description" : "Mala petición"
          }
        }
      }
    },
     "/roster/validate" : {
      "post" : {
        "tags" : [ "roster-api" ],
        "summary" : "Validate Roster",
        "description" : "Valida el Roster de Battlescribe contra Challonge, wahapedia.ru",
        "operationId" : "validateRoster",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ 
          {
          "name" : "roster_json",
          "in" : "formData",
          "description" : "Json Roster here.",
          "required" : true,
          "type" : "string",
          "default" : "{status:'VALIDATE', name:'Astra Millitarum-OT500-Primaris Psyker'}"
          }
        ],
        "responses" : {
          "200" : {
            "description" : "Resultado de la lista de tournaments[]",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Roster"
              }
            }
          },
          "400" : {
            "description" : "Mala petición"
          }
        }
      }
    },
    "/wahapedia/factions": {
      "get": {
        "tags": [
          "wahapedia-api"
        ],
        "summary": "Lista de Facciones y SubFacciones",
        "operationId": "getFactions",
        "description": "Obtiene la lista de Facciones y SubFacciones en Wahapedia.\n",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Resultado de exec del servicio",
            "schema": {
              "name" : "result",
              "type": "string"
            }
          },
          "400": {
            "description": "Mala petición"
          }
        }
      }
    },    
    "/challonge/tournaments": {
      "get": {
        "tags": [
          "challonge-tournaments"
        ],
        "summary": "Lista de Torneos",
        "operationId": "listTournaments",
        "description": "Obtiene la lista de Torneos en Challonge.\n",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Resultado de la lista de tournaments[]",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/tournament"
              }
            }
          },
          "400": {
            "description": "Mala petición"
          }
        }
      }
    },
    "/army/list": {
      "get": {
        "tags": [
          "army-api"
        ],
        "summary": "Lista de Armadas",
        "operationId": "listArmy",
        "description": "Obtiene la lista de Armadas en MTH40k DB.\n",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Resultado de la lista de Army[]",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Army"
              }
            }
          },
          "400": {
            "description": "Mala petición"
          }
        }
      }
    }
  },
  "definitions" : {
    "Faction" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "url" : {
          "type" : "string"
        }
      }
    },    
    "Roster" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        }
      }
    },
    "Army" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "example" : 1
        },
        "name" : {
          "type" : "string",
          "example" : "Xenos"
        },
        "code" : {
          "type" : "string",
          "example" : "XEN"
        },
        "factions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Faction"
          }
        }
      }
    },
    "tournament" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "example" : 7968393
        },
        "name" : {
          "type" : "string",
          "example" : "[ELI] Adeptus Astartes {OT500-V01}"
        },
        "url" : {
          "type" : "string"
        },
        "tournament_type" : {
          "type" : "string"
        },
        "game_name" : {
          "type" : "string"
        },
        "state" : {
          "type" : "string"
        },
        "event_id" : {
          "type" : "integer",
          "example" : 27010
        },
        "started_at" : {
          "type" : "string"
        },
        "completed_at": {
            "type": "string"
        }
      }
    }
  },
  "basePath": "/"
}